requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.8.2
Markdown==3.3.6
psycopg2==2.8.6
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
zipp==3.6.0
